mcp-servers/wolico/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wolico-mcp"
version = "2.0.0"
description = "MCP server per Wolico — ferie, staffing, economics, dipendenti"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
wolico-mcp = "server:main"
